Fixes global styles on sb
