+++ .storybook/main.ts
@@ -5,6 +5,7 @@ const config: StorybookConfig = {
     "@storybook/addon-links",
     "@storybook/addon-essentials",
     "@storybook/addon-interactions",
+    "@storybook/addon-styling",
   ],
   framework: {
     name: "@storybook/nextjs",
